Code restructing to move remeshing functions out of main adapter files
